#165590526 Notification email, endpoints validations and their tests

+++ server/controllers/accounts.js
@@ -1,6 +1,7 @@
+
 import moment from 'moment';
 import pool from '../config/db';
-import validation from '../../dummy/helpers/accounts';
+import validation from '../helpers/accounts';
 
 
 class Account {
@@ -9,49 +10,56 @@ class Account {
     try {
       const { error } = validation.AccountsValidation(req.body);
       if (error) {
+        const responseError = [];
+        error.details.map((e) => {
+          responseError.push({ message: e.message });
+        });
+
         return res.status(400).json({
           status: 400,
-          error: error.details[0].message,
+          error: responseError,
         });
       }
-      const getAccounts = 'SELECT * FROM accounts';
-      const { rows } = await pool.query(getAccounts);
-      const random = Math.floor(Math.random() * 10000000) + 100;
-      const accountnumber = parseInt(`4000${  random  }${rows + 1}`, 10);
-
-      const accountValues = {
-        accountNumber: accountnumber,
-        createdOn: moment().format('LL'),
-        owner: req.user.id,
-        type: req.body.type,
-        status: 'active',
-        balance: parseFloat(0),
-      };
       if (req.user.type !== 'client') {
         return res.status(403).json({
           status: 403,
           message: 'You are not allowed to perform this oparation!',
         });
       }
-      const queryText = 'INSERT INTO accounts (accountnumber, createdon, owner, type, status, balance) VALUES($1, $2, $3, $4, $5, $6)';
-      await pool.query(queryText, [accountValues.accountNumber,
-        accountValues.createdOn,
-        accountValues.owner,
-        accountValues.type,
-        accountValues.status,
-        accountValues.balance]);
+      if (req.user.type === 'client') {
+        const getAccounts = 'SELECT * FROM accounts';
+        const { rows } = await pool.query(getAccounts);
+        const random = Math.floor(Math.random() * 10000000) + 100;
+        const accountnumber = parseInt(`4000${  random  }${rows + 1}`, 10);
 
-      return res.status(201).json({
-        status: 201,
-        data: {
-          accountNumber: accountValues.accountNumber,
-          firstName: req.user.firstName,
-          lastName: req.user.lastName,
-          email: req.user.email,
-          type: accountValues.type,
-          openingBalance: accountValues.balance,
-        },
-      });
+        const accountValues = {
+          accountNumber: accountnumber,
+          createdOn: moment().format('LL'),
+          owner: req.user.id,
+          type: req.body.type,
+          status: 'active',
+          balance: parseFloat(0),
+        };
+        const queryText = 'INSERT INTO accounts (accountnumber, createdon, owner, type, status, balance) VALUES($1, $2, $3, $4, $5, $6)';
+        await pool.query(queryText, [accountValues.accountNumber,
+          accountValues.createdOn,
+          accountValues.owner,
+          accountValues.type,
+          accountValues.status,
+          accountValues.balance]);
+
+        return res.status(201).json({
+          status: 201,
+          data: {
+            accountNumber: accountValues.accountNumber,
+            firstName: req.user.firstName,
+            lastName: req.user.lastName,
+            email: req.user.email,
+            type: accountValues.type,
+            openingBalance: accountValues.balance,
+          },
+        });
+      }
     } catch (err) {
       return res.status(500).json({
         status: 500,
@@ -71,43 +79,56 @@ class Account {
         });
       }
       // verify if the account activate or deactive exist
-      const getAccount = 'SELECT * FROM accounts WHERE accountnumber = $1';
-      const enteredAcc = parseInt(req.params.accountNumber, 10);
-      const { rows } = await pool.query(getAccount, [enteredAcc]);
-      if (!rows[0]) {
-        return res.status(404).json({
-          status: 404,
-          message: 'The account you are trying to activate or deactivate do not exist',
-        });
-      }
-
       if ((isNaN(req.params.accountNumber))) {
         return res.status(400).json({
           status: 400,
           message: 'Sorry the account number do not exist or is not an integer',
         });
       }
-      if (req.user.type !== 'staff') {
+      if (req.user.isadmin !== true) {
         return res.status(403).json({
           status: 403,
           message: 'Sorry you are not Authorized to perform this oparation!',
         });
       }
-      const queryText = 'UPDATE accounts SET status = $1 WHERE accountnumber = $2';
-      const values = [req.body.status, enteredAcc];
-      await pool.query(queryText, values);
+      if (req.user.isadmin === true) {
+        const getAccount = 'SELECT * FROM accounts WHERE accountnumber = $1';
+        const enteredAcc = parseInt(req.params.accountNumber, 10);
+        const { rows } = await pool.query(getAccount, [enteredAcc]);
+        if (!rows[0]) {
+          return res.status(404).json({
+            status: 404,
+            message: 'The account you are trying to activate or deactivate do not exist',
+          });
+        }
+        if (rows[0].status === 'active' && req.body.status === 'active') {
+          return res.status(400).json({
+            status: 400,
+            message: 'The account is already ACTIVE',
+          });
+        }
+        if (rows[0].status === 'dormant' && req.body.status === 'dormant') {
+          return res.status(400).json({
+            status: 400,
+            message: 'The account is already DORMANT',
+          });
+        }
+        const queryText = 'UPDATE accounts SET status = $1 WHERE accountnumber = $2';
+        const values = [req.body.status, enteredAcc];
+        await pool.query(queryText, values);
 
-      return res.status(200).json({
-        status: 200,
-        data: {
-          accountData: rows[0].accountnumber,
-          status: rows[0].status,
-          ownerId: rows[0].owner,
-          email: rows[0].email,
-          accountBalance: rows[0].balance,
-        },
-        message: 'The account has been updated',
-      });
+        return res.status(200).json({
+          status: 200,
+          data: {
+            accountData: rows[0].accountnumber,
+            status: rows[0].status,
+            ownerId: rows[0].owner,
+            email: rows[0].email,
+            accountBalance: rows[0].balance,
+          },
+          message: `The account has been updated to ${req.body.status}`,
+        });
+      }
     } catch (err) {
       return res.status(500).json({
         status: 500,
@@ -116,32 +137,43 @@ class Account {
     }
   }
 
+
   // DELETE A BANK ACCOUNT
   static async deleteAccount(req, res) {
     try {
-      const getAccount = 'SELECT * FROM accounts WHERE accountnumber = $1';
-      const enteredAcc = parseInt(req.params.accountNumber, 10);
-      const { rows } = await pool.query(getAccount, [enteredAcc]);
-      if (!rows[0]) {
-        return res.status(404).json({
-          status: 404,
-          message: 'The account you are trying to Delete do not exist',
-        });
-      }
-
       if (req.user.type !== 'staff') {
         return res.status(403).json({
           status: 403,
           message: 'You are not allowed to perform this oparation!',
         });
       }
-      const queryText = 'DELETE FROM accounts WHERE accountnumber = $1';
-      await pool.query(queryText, [enteredAcc]);
+      if (req.user.type === 'staff') {
+        const getAccount = 'SELECT * FROM accounts WHERE accountnumber = $1';
+        const enteredAcc = parseInt(req.params.accountNumber, 10);
+        const { rows } = await pool.query(getAccount, [enteredAcc]);
+        if (!rows[0]) {
+          return res.status(404).json({
+            status: 404,
+            message: 'The account you are trying to Delete do not exist',
+          });
+        }
+        if (rows[0].balance > 0) {
+          return res.status(400).json({
+            status: 400,
+            message: `The account you are trying to delete has some amount on it and you can not delete an account with money, the amount is ${rows[0].balance}`,
+          });
+        }
 
-      return res.status(200).json({
-        status: 200,
-        message: 'The bank account has been deleted successfully',
-      });
+        if (rows[0].balance <= 0) {
+          const queryText = 'DELETE FROM accounts WHERE accountnumber = $1';
+          await pool.query(queryText, [enteredAcc]);
+
+          return res.status(200).json({
+            status: 200,
+            message: 'The bank account has been deleted successfully',
+          });
+        }
+      }
     } catch (err) {
       return res.status(500).json({
         status: 500,
@@ -159,19 +191,21 @@ class Account {
           message: 'You are not allowed to perform this oparation!',
         });
       }
-      const accDetailsQueryText = 'SELECT createdon, accountnumber, email, accounts.type, status, balance FROM accounts INNER JOIN users ON users.id = accounts.owner WHERE accountnumber = $1';
-      const { rows } = await pool.query(accDetailsQueryText,
-        [parseInt(req.params.accountNumber, 10)]);
-      if (!rows[0]) {
-        return res.status(404).json({
-          status: 404,
-          message: 'Sorry that Account does not exists',
+      if (req.user.type === 'client') {
+        const accDetailsQueryText = 'SELECT createdon, accountnumber, email, accounts.type, status, balance FROM accounts INNER JOIN users ON users.id = accounts.owner WHERE accountnumber = $1';
+        const { rows } = await pool.query(accDetailsQueryText,
+          [parseInt(req.params.accountNumber, 10)]);
+        if (!rows[0]) {
+          return res.status(404).json({
+            status: 404,
+            message: 'Sorry that Account does not exists',
+          });
+        }
+        return res.status(200).json({
+          status: 200,
+          data: rows[0],
         });
       }
-      return res.status(200).json({
-        status: 200,
-        data: rows[0],
-      });
     } catch (err) {
       return res.status(500).json({
         status: 500,
@@ -198,19 +232,21 @@ class Account {
           message: 'Sorry you are not authorized to perform this operation!',
         });
       }
-      const innnerJoinQueryText = 'SELECT createdon, accountnumber, accounts.type, status, balance FROM accounts INNER JOIN users ON users.id = accounts.owner WHERE email = $1';
-      const results = await pool.query(innnerJoinQueryText, [req.params.email.toLowerCase()]);
+      if (req.user.type === 'staff') {
+        const innnerJoinQueryText = 'SELECT createdon, accountnumber, accounts.type, status, balance FROM accounts INNER JOIN users ON users.id = accounts.owner WHERE email = $1';
+        const results = await pool.query(innnerJoinQueryText, [req.params.email.toLowerCase()]);
 
-      if (results.rows.length === 0) {
-        return res.status(404).json({
-          status: 404,
-          message: 'Sorry the user has no account!',
+        if (results.rows.length === 0) {
+          return res.status(404).json({
+            status: 404,
+            message: 'Sorry the user has no account!',
+          });
+        }
+        return res.status(200).json({
+          status: 200,
+          data: results.rows,
         });
       }
-      return res.status(200).json({
-        status: 200,
-        data: results.rows,
-      });
     } catch (err) {
       return res.status(500).json({
         status: 500,
@@ -223,7 +259,6 @@ class Account {
   static async getAllCounts(req, res) {
     try {
       if (req.user.type !== 'staff') {
-        // console.log(req.user);
         return res.status(403).json({
           status: 403,
           message: 'You are not allowed to perform this oparation!',
@@ -254,7 +289,6 @@ class Account {
   static async getAllAccountsByStatus(req, res) {
     try {
       if (req.user.type !== 'staff') {
-        // console.log(req.user);
         return res.status(403).json({
           status: 403,
           message: 'Sorry you are not Authorized to perform this oparation!',
@@ -263,7 +297,6 @@ class Account {
       const queryText = 'SELECT createdon, accountnumber, email, accounts.type, status, balance FROM accounts INNER JOIN users ON users.id = accounts.owner WHERE status = $1';
       const value = [req.query.status];
       const { rows } = await pool.query(queryText, value);
-      // console.log(rows);
       if (rows.length === 0) {
         return res.status(404).json({
           status: 404,
